+++ src/LDL/Framework/Base/Collection/Contracts/CollectionInterface.php
@@ -2,7 +2,7 @@
 
 namespace LDL\Framework\Base\Collection\Contracts;
 
-use LDL\Framework\Base\Exception\CollectionException;
+use LDL\Framework\Base\Collection\Exception\CollectionException;
 use LDL\Framework\Base\Contracts\ToArrayInterface;
 
 interface CollectionInterface extends \Countable, \Iterator, ToArrayInterface