+++ src/LDL/Framework/Base/Collection/Contracts/LockAppendInterface.php
@@ -1,0 +1,16 @@
+<?php declare(strict_types=1);
+
+namespace LDL\Framework\Base\Collection\Contracts;
+
+interface LockAppendInterface
+{
+    /**
+     * @return CollectionInterface
+     */
+    public function lockAppend(): CollectionInterface;
+
+    /**
+     * @return bool
+     */
+    public function isAppendLock() : bool;
+}