+++ src/LDL/Framework/Base/Collection/Exception/LockReplaceException.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace LDL\Framework\Base\Collection\Exception;
+
+use LDL\Framework\Base\Exception\LockingException;
+
+class LockReplaceException extends LockingException
+{
+
+}