+++ src/LDL/Framework/Base/Collection/Traits/AppendableInterfaceTrait.php
@@ -10,10 +10,26 @@ use LDL\Framework\Base\Collection\Contracts\CollectionInterface;
 
 trait AppendableInterfaceTrait
 {
+    /**
+     * @var callable|null
+     */
+    private $_tBeforeAppendCallback;
+
     //<editor-fold desc="AppendableInterface methods">
+    public function onBeforeAppend($item, $key = null): void
+    {
+        if(null === $this->_tBeforeAppendCallback){
+            return;
+        }
+
+        ($this->_tBeforeAppendCallback)($this, $item, $key);
+    }
+
     public function append($item, $key = null): CollectionInterface
     {
         $key = $key ?? $this->count;
+
+        $this->onBeforeAppend($item, $key);
 
         if(null === $this->first){
             $this->first = $key;
@@ -23,15 +39,6 @@ trait AppendableInterfaceTrait
 
         $this->items[$key] = $item;
         $this->count++;
-
-        return $this;
-    }
-
-    public function appendMany(iterable $items, bool $useKey=false) : CollectionInterface
-    {
-        foreach ($items as $key => $value) {
-            $this->append($value, $useKey ? $key : null);
-        }
 
         return $this;
     }