+++ src/LDL/Framework/Base/Collection/Traits/RemovableInterfaceTrait.php
@@ -11,12 +11,31 @@ use LDL\Framework\Base\Collection\Exception\RemoveException;
 
 trait RemovableInterfaceTrait
 {
+    /**
+     * @var callable|null
+     */
+    private $_tBeforeRemoveCallback;
+
     //<editor-fold desc="RemovableInterface methods">
+    public function onBeforeRemove($item, $key): void
+    {
+        if(null === $this->_tBeforeRemoveCallback){
+            return;
+        }
+
+        ($this->_tBeforeRemoveCallback)($this, $item, $key);
+    }
+
     public function remove($key): CollectionInterface
     {
-        if(array_key_exists($this->items, $key)){
+        $exists = array_key_exists($key, $this->items);
+
+        $this->onBeforeRemove($exists ? $this->items[$key] : null, $key);
+
+        if($exists){
             unset($this->items[$key]);
             $this->count--;
+            return $this;
         }
 
         throw new RemoveException("Item with key: $key does not exists");